Now interface looks more complete. Still need rework some places with bug potential.
