+++ Project1/Debug/Project1.log
@@ -1,4 +1,3 @@
 ﻿  main.cpp
-G:\Learning\VS2019\GitTest4Studs\Project1\main.cpp(119,23): warning C4018: '<': signed/unsigned mismatch
-G:\Learning\VS2019\GitTest4Studs\Project1\main.cpp(262): warning C4715: 'interfaceMenu': not all control paths return a value
+G:\Learning\VS2019\GitTest4Studs\Project1\main.cpp(44,23): warning C4018: '<': signed/unsigned mismatch
   Project1.vcxproj -> G:\Learning\VS2019\GitTest4Studs\Project1\Debug\Project1.exe
